Move postgres configuration to separate crate
