+++ Cargo.toml
@@ -2,4 +2,5 @@
 members = [
   "crates/brace-cms",
   "crates/brace-cms-log",
+  "crates/brace-cms-store-postgres",
 ]
