+++ crates/brace-cms-store-postgres/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "brace-cms-store-postgres"
+version = "0.1.0"
+authors = ["Daniel Balcomb <daniel.balcomb@gmail.com>"]
+description = "The postgres data store integration."
+repository = "https://github.com/brace-rs/brace-cms"
+license = "MIT OR Apache-2.0"
+edition = "2018"
+
+[dependencies]
+brace-config = { git = "https://github.com/brace-rs/brace-config", rev = "da563f455a8d4e98d90dc2f27c639ea789f078c2", default_features = false }
+brace-data-store-postgres = { git = "https://github.com/brace-rs/brace-data", rev = "50592b806e748d922cd0461abbdcb1a2bb0e2062" }
+brace-web-core = { git = "https://github.com/brace-rs/brace-web", rev = "dd0cee2e916553594acb8270632b5727474db01f" }
+futures = "0.3"
