+++ crates/brace-cms-store-postgres/src/lib.rs
@@ -1,0 +1,5 @@
+pub use brace_data_store_postgres::*;
+
+pub use self::config::configure;
+
+pub mod config;
