+++ crates/brace-cms/Cargo.toml
@@ -17,10 +17,9 @@ dev = ["listenfd"]
 [dependencies]
 actix-rt = "1.0"
 brace-cms-log = { path = "../brace-cms-log" }
+brace-cms-store-postgres = { path = "../brace-cms-store-postgres" }
 brace-config = { git = "https://github.com/brace-rs/brace-config", rev = "da563f455a8d4e98d90dc2f27c639ea789f078c2", features = ["toml"], default_features = false }
-brace-data-store-postgres = { git = "https://github.com/brace-rs/brace-data", rev = "50592b806e748d922cd0461abbdcb1a2bb0e2062" }
 brace-web = { git = "https://github.com/brace-rs/brace-web", rev = "dd0cee2e916553594acb8270632b5727474db01f" }
-futures = "0.3"
 listenfd = { version = "0.3", optional = true }
 
 [dev-dependencies]
