+++ crates/brace-cms/src/lib/lib.rs
@@ -1,29 +1,18 @@
 use std::io;
 use std::net::Ipv4Addr;
 
-use futures::future::TryFutureExt;
-
 use brace_config::Config;
-use brace_data_store_postgres::{Config as PostgresConfig, Error as PostgresError, Postgres};
 use brace_web::core::middleware::Logger;
-use brace_web::core::{web, App, Error, HttpResponse, HttpServer};
+use brace_web::core::{web, App, HttpResponse, HttpServer};
+
+pub mod store {
+    pub mod postgres {
+        pub use brace_cms_store_postgres::*;
+    }
+}
 
 async fn index() -> HttpResponse {
     HttpResponse::Ok().body("Hello world")
-}
-
-async fn postgres_info(postgres: web::Data<Postgres>) -> Result<HttpResponse, Error> {
-    let version = postgres_version(&postgres).map_err(|_| ()).await?;
-
-    Ok(HttpResponse::Ok().body(format!("postgres version: {}", version)))
-}
-
-async fn postgres_version(postgres: &Postgres) -> Result<String, PostgresError> {
-    let conn = postgres.connect().await?;
-    let row = conn.query_one("SELECT version()", &[]).await?;
-    let res = row.get::<_, String>(0);
-
-    Ok(res)
 }
 
 pub async fn server(config: Config) -> io::Result<()> {
@@ -36,33 +25,13 @@ pub async fn server(config: Config) -> io::Result<()> {
     let addr = format!("{}:{}", host, port);
     let format = r#"%a "%r" %s %b "%{Referer}i" "%{User-Agent}i" %T"#;
 
-    let mut postgres_config = PostgresConfig::new();
-
-    postgres_config.host(
-        &config
-            .get::<_, String>("postgres.host")
-            .unwrap_or_else(|_| String::from("localhost")),
-    );
-    postgres_config.port(config.get::<_, u16>("postgres.port").unwrap_or(5432));
-    postgres_config.user(
-        &config
-            .get::<_, String>("postgres.user")
-            .unwrap_or_else(|_| String::from("postgres")),
-    );
-    postgres_config.password(
-        &config
-            .get::<_, String>("postgres.pass")
-            .unwrap_or_else(|_| String::from("postgres")),
-    );
-
-    let postgres = Postgres::new(postgres_config).await.unwrap();
+    let postgres = brace_cms_store_postgres::configure(&config).await;
 
     let mut server = HttpServer::new(move || {
         App::new()
             .wrap(Logger::new(format))
-            .data(postgres.clone())
+            .configure(postgres.clone())
             .route("/", web::get().to(index))
-            .route("/postgres", web::get().to(postgres_info))
     });
 
     #[cfg(feature = "dev")]
